refactor, css, bugs
everything broke, started partial view.
basic css that needs to be redone.
started remake of routing and url design
+ Mob programming with @45an @boskojr
