+++ Controllers/HomeController.cs
@@ -7,15 +7,18 @@ namespace _4chanForum.Controllers
     public class HomeController : Controller
     {
         private readonly ILogger<HomeController> _logger;
+        private readonly DataContext _context;
 
-        public HomeController(ILogger<HomeController> logger)
+        public HomeController(ILogger<HomeController> logger, DataContext context)
         {
             _logger = logger;
+            _context = context;
         }
 
         public IActionResult Index()
         {
-            return View();
+            var topics = _context.Topics.ToList();
+            return View(topics);
         }
 
         public IActionResult Privacy()
