+++ Models/ReplyModel.cs
@@ -6,5 +6,6 @@
         public string? Title { get; set; }
         public string? Text { get; set; }
         public int ThreadId { get; set; }
+        public bool isPinned { get; set; } = false;
     }
 }
