+++ Models/TopicModel.cs
@@ -5,5 +5,6 @@
         public int Id { get; set; }
         public string? Title { get; set; }
         public string? Description { get; set; }
+        public bool isPinned { get; set; } = false;
     }
 }
