+++ Views/Shared/_Layout.cshtml
@@ -19,12 +19,7 @@
                 </button>
                 <div class="navbar-collapse collapse d-sm-inline-flex justify-content-between">
                     <ul class="navbar-nav flex-grow-1">
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Index">Home</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Forum" asp-action="Topics">Forum</a>
-                        </li>
+                        <partial name="_TopicListPartial" class="nav-item"/>
                     </ul>
                 </div>
             </div>
